Update list layout for relearn
The previous list layout reimplemented the list.html layout from the
learn theme. The relearn theme has completely reimplemented these
templates, so this old template no longer worked.

The list template was special in that it printed a summary list of
subpages. This has been moved to the content partial. This partial
is used for both single and list pages, so the partial checks if
there are any subpages, and if so it creates the table at the bottom.