Fix links that hugo/relearn is complaining about
The relearn theme is doing some extra rerendering tricks when building
pages, and part of that is checking links. This change fixes most of
the problems raised. (Most of the errors were caused by hugo apparently
not being able to trace relative links up a directory.)

Also downgrading the check from error to warning. As much as I would
like to make sure all links are live, there are still some that I
cannot find a way to work around.