Turn off child summary on landing page
In the setup for the learn theme, the child summary was in the
list layout, and this layout is not used in the main landing page.
In relearn, the child summary code is on all pages, so it has to
be suppressed on the landing page (which is currently the only
page with children that should not show them).